feat: use modal to create and update share class
